+++ frontend/tsconfig.app.json
@@ -23,5 +23,9 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src"]
+  "include": [
+    "src",
+    "../admin-dashboard/src/pages/AdminLogin.tsx",
+    "../admin-dashboard/src/services"
+  ]
 }
